LPS-85293 users-admin-web - moves OrganizationScreenNavigationDisplayContext.java to a more semantic package
