+++ modules/apps/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/internal/display/context/OrganizationScreenNavigationDisplayContext.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.users.admin.web.internal.frontend.taglib.servlet.taglib;
+package com.liferay.users.admin.web.internal.display.context;
 
 import com.liferay.portal.kernel.model.Organization;
